requirements.txt
websocket-client
streamlit
pandas
sqlalchemy
psycopg2-binary   # only if using Postgres; skip if SQLite
python-dotenv
vaderSentiment
praw
